separated code and added some more types
