+++ src/dataProvider/index.ts
@@ -1,5 +1,6 @@
-import { fetchUtils, DataProvider, Identifier } from 'ra-core';
-
+import { fetchUtils, DataProvider } from 'ra-core';
+import { PrimaryKey, PostgRestOperator, getPrimaryKey, parseFilters, 
+         getOrderBy, dataWithId, getQuery, getKeyData, encodeId } from "./utils";
 /**
  * Maps react-admin queries to a postgrest REST API
  *
@@ -35,144 +36,11 @@ import { fetchUtils, DataProvider, Identifier } from 'ra-core';
  *
  * export default App;
  */
-function parseFilters(filter, defaultListOp) {
-  let result = {};
-  Object.keys(filter).forEach(function (key) {
-    // key: the name of the object key
 
-    const splitKey = key.split('@');
-    const operation = splitKey.length == 2 ? splitKey[1] : defaultListOp;
-
-    let values;
-    if (operation.includes('like')) {
-      // we split the search term in words
-      values = filter[key].trim().split(' ');
-    } else {
-      values = [filter[key]];
-    }
-
-    values.forEach(value => {
-      // if operator is intentionally blank, rpc syntax 
-      let op = operation.includes('like') ? `${operation}.*${value}*` : operation.length == 0 ? `${value}` : `${operation}.${value}`;
-
-      if (result[splitKey[0]] === undefined) {
-        // first operator for the key, we add it to the dict
-        result[splitKey[0]] = op;
-      }
-      else
-      {
-        if (!Array.isArray(result[splitKey[0]])) {
-          // second operator, we transform to an array
-          result[splitKey[0]] = [result[splitKey[0]], op]
-        } else {
-          // third and subsequent, we add to array
-          result[splitKey[0]].push(op);
-        }
-      }
-    });
-
-  });
-
-  return result;
-}
-
-// compound keys capability
-type PrimaryKey = Array<string>;
-
-const getPrimaryKey = (resource : string, primaryKeys: Map<string, PrimaryKey>) => {
-  return primaryKeys.get(resource) || ['id'];
-}
-
-const decodeId = (id: Identifier, primaryKey: PrimaryKey): string[] => {
-  if (isCompoundKey(primaryKey)) {
-    return JSON.parse(id.toString());
-  } else {
-    return [id.toString()];
-  }
-}
-
-const encodeId = (data: any, primaryKey: PrimaryKey): Identifier => {
-  if (isCompoundKey(primaryKey)) {
-    return JSON.stringify(primaryKey.map(key => data[key]));
-  } else {
-    return data[primaryKey[0]];
-  }
-}
-
-const dataWithId = (data: any, primaryKey: PrimaryKey) => {
-  if (primaryKey === ['id']) {
-    return data;
-  }
-
-  return Object.assign(data, {
-    id: encodeId(data, primaryKey)
-  });
-}
-
-const isCompoundKey = (primaryKey: PrimaryKey) : Boolean => {
-  return primaryKey.length > 1;
-}
-
-const getQuery = (primaryKey : PrimaryKey, ids: Identifier | Array<Identifier>, resource: string) : string => {
-  if (Array.isArray(ids) && ids.length > 1) {
-    // no standardized query with multiple ids possible for rpc endpoints which are api-exposed database functions
-    if (resource.startsWith('rpc/')) {
-      console.error('PostgREST\'s rpc endpoints are not intended to be handled as views. Therefore, no query generation for multiple key values implemented!');
-
-      return ;
-    }
-
-    if (isCompoundKey(primaryKey)) {
-      return `or=(
-          ${ids.map(id => {
-                const primaryKeyParams = decodeId(id, primaryKey);
-                return `and(${primaryKey.map((key, i) => `${key}.eq.${primaryKeyParams[i]}`).join(',')})`;
-              })
-            }
-        )`;
-      } else {
-        return new URLSearchParams({ [primaryKey[0]]: `in.(${ids.join(',')})` }).toString();
-      }
-  } else {
-    // if ids is one Identifier
-    const id : Identifier = ids.toString();
-    const primaryKeyParams = decodeId(id, primaryKey);
-
-    if (isCompoundKey(primaryKey)) {
-      if (resource.startsWith('rpc/'))
-        return `${primaryKey.map((key : string, i: any) => `${key}=${primaryKeyParams[i]}`).join('&')}`;
-      else
-        return `and=(${primaryKey.map((key : string, i: any) => `${key}.eq.${primaryKeyParams[i]}`).join(',')})`;
-    } else {
-      return new URLSearchParams([[primaryKey[0], `eq.${id}`]]).toString();
-    }
-  }
-}
-
-const getKeyData = (primaryKey : PrimaryKey, data: object) : object => {
-  if (isCompoundKey(primaryKey)) {
-    return primaryKey.reduce(
-      (keyData, key) => ({
-        ...keyData,
-          [key]: data[key]
-        }), 
-      {});
-  } else {
-    return { [primaryKey[0]]: data[primaryKey[0]] };
-  }
-}
-
-const getOrderBy = (field : string, order: string, primaryKey : PrimaryKey) => {
-  if (field == 'id') {
-    return primaryKey.map(key => (`${key}.${order.toLowerCase()}`)).join(',');
-  } else {
-    return `${field}.${order.toLowerCase()}`;
-  }
-};
 
 const defaultPrimaryKeys = new Map<string, PrimaryKey>();
 
-export default (apiUrl, httpClient = fetchUtils.fetchJson, defaultListOp = 'eq', 
+export default (apiUrl : string, httpClient = fetchUtils.fetchJson, defaultListOp : PostgRestOperator = 'eq', 
                 primaryKeys: Map<string, PrimaryKey> = defaultPrimaryKeys): DataProvider => ({
   getList: (resource, params) => {
     const primaryKey = getPrimaryKey(resource, primaryKeys);
